Refactor arborescence projet
