+++ src/app/components/app/app.component.scss
