+++ src/app/components/app/app.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { Pokemon } from "./pokemon";
-import { POKEMONS } from "./mock-pokemon-list";
+import { Pokemon } from "../../models/pokemon";
+import { POKEMONS } from "../../data/mock-pokemon-list";
 
 @Component({
   selector: 'app-root',
