+++ src/app/data/mock-pokemon-list.ts
@@ -1,4 +1,4 @@
-import { Pokemon } from './pokemon';
+import { Pokemon } from '../models/pokemon';
 
 export const POKEMONS: Pokemon[] = [
   {
